==> src/ccc/04_gene_regulatory_networks.Rmd <==
---
title: "04_gene_regulatory_networks"
author: "Tabea M. Soelter"
date: '2023-10-02'
output: html_document
---
**Determination of signaling mediators using gene regulatory networks from MultiNicheNet**

__Goal__: Identify all potential signaling mediators (receptor -> target gene) by determining neighbors of all receptors and predicting the most likely mediator through shortest path calculations using Dijkstra's algorithm. This is done only for interactions that validate across datasets.
  
__Reproducibility__:
* GitHub: lasseignelab/230313_TS_CCCinHumanAD
* Docker: tsoelter/rstudio_ccc_ad
    * Version: 1.0.5
* HPC: Yes
  * Resources: long partition (150hrs), 6 CPUs, 65GB per CPU

__Data__:
*MultiNicheNet outputs*
* Name: geo_multinichenet_output.rds, gse_multinichenet_output.rds
* Location: data/ccc/
 
__Analysis Plan__:
* Load necessary packages 
* Load data
* Prepare data by filtering (as performed previously)
* Grab LRTs that overlap between datasets
* Extract signaling GRNs from MultiNicheNet
  * Plot and save
* Create igraph objects for each LRT
  * Wrapper function
  * For both datasets
* Calculate network properties
  * Receptor neighbors = all signaling mediators
  * Shortest path = top mediator
* Save outputs and objects
  
# Load packages
```{r}
# Time tracking
ptm <- proc.time()

# Setting seed
set.seed(42)

# Setting locale
options(encoding = "UTF-8")

# Load packages
suppressPackageStartupMessages({
  library(multinichenetr)
  library(nichenetr)
  library(igraph)
  library(tidyverse)
  library(here)
  library(styler)
  library(lintr)
})

source(here("src", "functions_CCCinAD_tms.R"))
```

# Load data
```{r}
geo_multinichenet_output <- readRDS(here(
  "data",
  "ccc",
  "geo_multinichenet_output_final.rds"
))

gse_multinichenet_output <- readRDS(here(
  "data",
  "ccc",
  "gse_multinichenet_output_final.rds"
))

multinichenet_objects <- tibble::lst(
  geo_multinichenet_output,
  gse_multinichenet_output
)
```

# NicheNet-v2 prior
* downloaded here: https://zenodo.org/record/7074291
```{r}
lr_network <- readRDS(here(
  "data",
  "ccc",
  "nichenet_v2_prior",
  "lr_network_human_21122021.rds"
)) %>%
  dplyr::rename(ligand = from, receptor = to) %>%
  distinct(ligand, receptor) %>%
  mutate(ligand = make.names(ligand), receptor = make.names(receptor))

ligand_target_matrix <- readRDS(here(
  "data",
  "ccc",
  "nichenet_v2_prior",
  "ligand_target_matrix_nsga2r_final.rds"
))

# format row and column names
colnames(ligand_target_matrix) <- colnames(ligand_target_matrix) %>%
  make.names()
rownames(ligand_target_matrix) <- rownames(ligand_target_matrix) %>%
  make.names()
```

# MultiNicheNet GRN prior
* All of the priors were downloaded from the MultiNicheNet zenodo, as described in the MultiNicheNet vignette (accesssed 231002).
```{r}
sig_network <- readRDS(
  here(
    "data",
    "ccc",
    "nichenet_grn",
    "signaling_network_human_21122021.rds"
  )
) %>%
  mutate(
    from = make.names(from),
    to = make.names(to)
  )

gr_network <- readRDS(
  here(
    "data",
    "ccc",
    "nichenet_grn",
    "gr_network_human_21122021.rds"
  )
) %>%
  mutate(
    from = make.names(from),
    to = make.names(to)
  )

ligand_tf_matrix <- readRDS(
  here(
    "data",
    "ccc",
    "nichenet_grn",
    "ligand_tf_matrix_nsga2r_final.rds"
  )
)
colnames(ligand_tf_matrix) <- colnames(ligand_tf_matrix) %>%
  make.names()
rownames(ligand_tf_matrix) <- rownames(ligand_tf_matrix) %>%
  make.names()

weighted_networks <- readRDS(
  here(
    "data",
    "ccc",
    "nichenet_grn",
    "weighted_networks_nsga2r_final.rds"
  )
)
weighted_networks$lr_sig <- weighted_networks$lr_sig %>%
  mutate(
    from = make.names(from),
    to = make.names(to)
  )
weighted_networks$gr <- weighted_networks$gr %>%
  mutate(
    from = make.names(from),
    to = make.names(to)
  )
```

# Set filtering parameters
* NOTE: These are the same as in previous analyses.
```{r}
top_n_target <- 250

receiver_oi <- c("Excitatory.Neurons", "Inhibitory.Neurons")

sender_oi <- c("Astrocytes", "Microglia", "OPCs", "Oligodendrocytes")

contrast_tbl <- tibble(
  contrast = c("AD-CTRL", "CTRL-AD"),
  group = c("AD", "CTRL")
)
```

# Filter MultiNicheNet outputs
```{r}
filtered_objects <- filter_multinichenet(
  object_list = multinichenet_objects,
  sender_oi = sender_oi,
  receiver_oi = receiver_oi,
  contrast_tbl = contrast_tbl
)

list2env(filtered_objects, globalenv())
```

# Get overlapping LRTs
```{r}
# get LRTs per dataset
geo_id_target <- geo_lr_target_prior_cor_filtered$id_target %>%
  unique()

gse_id_target <- gse_lr_target_prior_cor_filtered$id_target %>%
  unique()

# get overlapping LRTs
overlap <- intersect(geo_id_target, gse_id_target)
length(overlap) # 7

# subset based on overlap for both datasets
geo_ccc <- geo_lr_target_prior_cor_filtered %>%
  dplyr::select(sender, receiver, ligand, target, receptor, id_target) %>%
  filter(id_target %in% overlap) %>%
  mutate(origin = "geo")

gse_ccc <- gse_lr_target_prior_cor_filtered %>%
  dplyr::select(sender, receiver, ligand, target, receptor, id_target) %>%
  filter(id_target %in% overlap) %>%
  mutate(origin = "gse")

# combine
ccc_combined <- rbind(geo_ccc, gse_ccc) %>%
  select(sender, receiver, ligand, receptor, target, id_target) %>%
  unique()
```

# Make igraph objects
* I am also plotting all the sub-networks and saving to specified directory for each dataset
```{r}
igraph_objects_list <- signaling_igraph(
  lrt_filtered_list = filtered_objects,
  overlap = overlap,
  ccc_combined = ccc_combined,
  ligand_tf_matrix = ligand_tf_matrix,
  weighted_networks = weighted_networks,
  lr_network = lr_network,
  sig_network = sig_network,
  gr_network = gr_network,
  plots = "results/intermediate_outputs/"
)

list2env(igraph_objects_list, globalenv())
```

# Calculate network properties
* I calculated many network topology metrics, however the ones highlighted in the manuscript are:
  * neighbors: all nodes directly connected to our receptor node. Must be outgoing from the receptor.
  * shortest path: the path from receptor to target gene that is the shortest. This represents our top or most-likely mediator.
  
* NOTE: This chunk produces warnings due to missing information to calculate shortest path. That is expected.
```{r include=FALSE}
geo_network_proporties <- network_topology(geo_igraph_objects)

gse_network_proporties <- network_topology(gse_igraph_objects)
```

# Save outputs
```{r}
saveRDS(geo_network_proporties, file = here(
  "results",
  "final_outputs",
  "geo",
  "signaling_grn_properties_final.rds"
))

saveRDS(gse_network_proporties, file = here(
  "results",
  "final_outputs",
  "gse",
  "signaling_grn_properties_final.rds"
))
```

# Save igraph objects
```{r}
saveRDS(geo_igraph_objects, file = here(
  "data",
  "ccc",
  "geo_signaling_igraph_objects_final.rds"
))

saveRDS(gse_igraph_objects, file = here(
  "data",
  "ccc",
  "gse_signaling_igraph_objects_final.rds"
))
```

# Session info
```{r}
sessionInfo() # see output below
```
R version 4.1.3 (2022-03-10)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 20.04.6 LTS

Matrix products: default
BLAS/LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.8.so

locale:
[1] C

attached base packages:
[1] stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
 [1] lintr_3.0.2          styler_1.9.1         here_1.0.1           lubridate_1.9.2     
 [5] forcats_1.0.0        stringr_1.5.0        dplyr_1.1.3          purrr_1.0.2         
 [9] readr_2.1.4          tidyr_1.3.0          tibble_3.2.1         ggplot2_3.4.2       
[13] tidyverse_2.0.0      igraph_1.5.1         nichenetr_2.0.0      multinichenetr_1.0.0

loaded via a namespace (and not attached):
  [1] rsvd_1.0.5                  Hmisc_5.1-0                 ica_1.0-3                  
  [4] ps_1.7.5                    class_7.3-20                rprojroot_2.0.3            
  [7] foreach_1.5.2               lmtest_0.9-40               crayon_1.5.2               
 [10] rhdf5filters_1.6.0          rbibutils_2.2.13            MASS_7.3-55                
 [13] nlme_3.1-155                backports_1.4.1             sva_3.42.0                 
 [16] rlang_1.1.1                 XVector_0.34.0              caret_6.0-94               
 [19] ROCR_1.0-11                 irlba_2.3.5.1               callr_3.7.3                
 [22] nloptr_2.0.3                limma_3.50.3                scater_1.22.0              
 [25] BiocParallel_1.28.3         rjson_0.2.21                bit64_4.0.5                
 [28] glue_1.6.2                  sctransform_0.3.5.9006      processx_3.8.1             
 [31] pbkrtest_0.5.2              parallel_4.1.3              vipor_0.4.5                
 [34] spatstat.sparse_3.0-2       AnnotationDbi_1.56.2        BiocGenerics_0.40.0        
 [37] UpSetR_1.4.0                muscat_1.8.2                spatstat.geom_3.2-2        
 [40] tidyselect_1.2.0            SummarizedExperiment_1.24.0 SeuratObject_4.1.3         
 [43] fitdistrplus_1.1-11         variancePartition_1.24.1    XML_3.99-0.14              
 [46] zoo_1.8-12                  ggpubr_0.6.0                xtable_1.8-4               
 [49] magrittr_2.0.3              evaluate_0.21               Rdpack_2.4                 
 [52] scuttle_1.4.0               cli_3.6.1                   zlibbioc_1.40.0            
 [55] rstudioapi_0.14             miniUI_0.1.1.1              sp_2.0-0                   
 [58] rpart_4.1.16                aod_1.3.2                   locfdr_1.1-8               
 [61] shiny_1.7.4.1               BiocSingular_1.10.0         xfun_0.40                  
 [64] clue_0.3-64                 cluster_2.1.2               caTools_1.18.2             
 [67] tidygraph_1.2.3             KEGGREST_1.34.0             ggrepel_0.9.3              
 [70] listenv_0.9.0               Biostrings_2.62.0           png_0.1-8                  
 [73] future_1.33.0               ipred_0.9-14                withr_2.5.0                
 [76] bitops_1.0-7                ggforce_0.4.1               plyr_1.8.8                 
 [79] hardhat_1.3.0               dqrng_0.3.0                 e1071_1.7-13               
 [82] pROC_1.18.2                 coda_0.19-4                 pillar_1.9.0               
 [85] gplots_3.1.3                GlobalOptions_0.1.2         cachem_1.0.8               
 [88] GetoptLong_1.0.5            DelayedMatrixStats_1.16.0   vctrs_0.6.3                
 [91] ellipsis_0.3.2              generics_0.1.3              lava_1.7.2.1               
 [94] tools_4.1.3                 foreign_0.8-82              beeswarm_0.4.0             
 [97] munsell_0.5.0               tweenr_2.0.2                emmeans_1.8.6              
[100] proxy_0.4-27                DelayedArray_0.20.0         fastmap_1.1.1              
[103] compiler_4.1.3              abind_1.4-5                 httpuv_1.6.11              
[106] plotly_4.10.2               GenomeInfoDbData_1.2.7      prodlim_2023.03.31         
[109] gridExtra_2.3               glmmTMB_1.1.7               edgeR_3.36.0               
[112] lattice_0.20-45             ggnewscale_0.4.9            deldir_1.0-9               
[115] visNetwork_2.1.2            utf8_1.2.3                  later_1.3.1                
[118] recipes_1.0.6               jsonlite_1.8.7              scales_1.2.1               
[121] ScaledMatrix_1.2.0          carData_3.0-5               pbapply_1.7-2              
[124] sparseMatrixStats_1.6.0     estimability_1.4.1          genefilter_1.76.0          
[127] lazyeval_0.2.2              promises_1.2.0.1            car_3.1-2                  
[130] doParallel_1.0.17           R.utils_2.12.2              goftest_1.2-3              
[133] spatstat.utils_3.0-3        reticulate_1.30             checkmate_2.2.0            
[136] cyclocomp_1.1.0             rmarkdown_2.25              cowplot_1.1.1              
[139] blme_1.0-5                  statmod_1.5.0               Rtsne_0.16                 
[142] Biobase_2.54.0              uwot_0.1.16                 HDF5Array_1.22.1           
[145] yaml_2.3.7                  survival_3.3-1              numDeriv_2016.8-1.1        
[148] htmltools_0.5.6             memoise_2.0.1               Seurat_4.3.0.9011          
[151] locfit_1.5-9.8              graphlayouts_1.0.0          IRanges_2.28.0             
[154] viridisLite_0.4.2           digest_0.6.33               RhpcBLASctl_0.23-42        
[157] mime_0.12                   RSQLite_2.3.1               future.apply_1.11.0        
[160] remotes_2.4.2               data.table_1.14.8           blob_1.2.4                 
[163] R.oo_1.25.0                 S4Vectors_0.32.4            DiagrammeR_1.0.10          
[166] labeling_0.4.2              splines_4.1.3               Formula_1.2-5              
[169] Rhdf5lib_1.16.0             RCurl_1.98-1.12             broom_1.0.5                
[172] hms_1.1.3                   rhdf5_2.38.1                DropletUtils_1.14.2        
[175] colorspace_2.1-0            base64enc_0.1-3             ggbeeswarm_0.7.2           
[178] GenomicRanges_1.46.1        shape_1.4.6                 nnet_7.3-17                
[181] Rcpp_1.0.11                 RANN_2.6.1                  mvtnorm_1.2-2              
[184] circlize_0.4.15             fansi_1.0.4                 tzdb_0.4.0                 
[187] parallelly_1.36.0           ModelMetrics_1.2.2.2        R6_2.5.1                   
[190] grid_4.1.3                  factoextra_1.0.7            ggridges_0.5.4             
[193] lifecycle_1.0.3             bluster_1.4.0               ggsignif_0.6.4             
[196] minqa_1.2.5                 leiden_0.4.3                Matrix_1.5-4               
[199] desc_1.4.2                  RcppAnnoy_0.0.21            RColorBrewer_1.1-3         
[202] iterators_1.0.14            spatstat.explore_3.2-1      TMB_1.9.4                  
[205] gower_1.0.1                 R.cache_0.16.0              htmlwidgets_1.6.2          
[208] beachmat_2.10.0             polyclip_1.10-4             shadowtext_0.1.2           
[211] timechange_0.2.0            mgcv_1.8-39                 ComplexHeatmap_2.10.0      
[214] globals_0.16.2              htmlTable_2.4.1             patchwork_1.1.2            
[217] spatstat.random_3.1-5       progressr_0.13.0            codetools_0.2-18           
[220] matrixStats_1.0.0           metapod_1.2.0               randomForest_4.7-1.1       
[223] gtools_3.9.4                prettyunits_1.2.0           SingleCellExperiment_1.16.0
[226] R.methodsS3_1.8.2           GenomeInfoDb_1.30.1         gtable_0.3.3               
[229] DBI_1.1.3                   stats4_4.1.3                tensor_1.5                 
[232] httr_1.4.7                  KernSmooth_2.23-20          stringi_1.7.12             
[235] progress_1.2.2              reshape2_1.4.4              farver_2.1.1               
[238] annotate_1.72.0             viridis_0.6.3               rex_1.2.1                  
[241] fdrtool_1.2.17              timeDate_4022.108           xml2_1.3.5                 
[244] boot_1.3-28                 BiocNeighbors_1.12.0        lme4_1.1-33                
[247] geneplotter_1.72.0          scattermore_1.2             scran_1.22.1               
[250] DESeq2_1.34.0               bit_4.0.5                   MatrixGenerics_1.6.0       
[253] spatstat.data_3.0-1         ggraph_2.1.0                pkgconfig_2.0.3            
[256] lmerTest_3.1-3              rstatix_0.7.2               knitr_1.44 

# Time tracking
```{r}
fptm <- proc.time() - ptm

# convert to minutes
(fptm[3] / 60)

# Time elapsed: 15 minutes
```

# Reproducibility
```{r}
# styler
style_file(here(
  "src",
  "ccc",
  "04_gene_regulatory_networks.Rmd"
))

# lintr
lint(filename = here(
  "src",
  "ccc",
  "04_gene_regulatory_networks.Rmd"
))
```
